2025.11.17- 김유리 주문등록 계산 수정

+++ src/main/java/mes/domain/entity/actasEntity/TB_DA006W.java
@@ -124,7 +124,7 @@ public class TB_DA006W {
 
     // 30 개당이윤
     @Column(name = "pereyunamt")
-    private Long pereyunamt;
+    private BigDecimal pereyunamt;
 
     // 31 이윤율
     @Column(name = "eyunyul")
